Add an abstract MessageCompiler class
