+++ lib/dry/validation/error_compiler.rb
@@ -1,50 +1,14 @@
-require 'dry/validation/message'
-require 'dry/validation/message_set'
+require 'dry/validation/message_compiler'
 
 module Dry
   module Validation
-    class ErrorCompiler
-      attr_reader :messages, :options, :locale
-
-      EMPTY_ARRAY = [].freeze
-
-      def initialize(messages, options = {})
-        @messages = messages
-        @options = Hash[options]
-        @full = @options.fetch(:full, false)
-        @locale = @options.fetch(:locale, :en)
-      end
-
+    class ErrorCompiler < MessageCompiler
       def message_type
         :failure
       end
 
       def message_class
         Message
-      end
-
-      def full?
-        @full
-      end
-
-      def call(ast)
-        MessageSet[ast.map { |node| visit(node) }]
-      end
-
-      def with(new_options)
-        self.class.new(messages, options.merge(new_options))
-      end
-
-      def visit(node, *args)
-        __send__(:"visit_#{node[0]}", node[1], *args)
-      end
-
-      def visit_schema(node, opts = {})
-        visit(node)
-      end
-
-      def visit_set(node)
-        call(node)
       end
 
       def visit_error(node, opts = {})
@@ -76,138 +40,13 @@ module Dry
         visit(other, opts.merge(input: input))
       end
 
-      def visit_implication(node, *args)
-        _, right = node
-        visit(right, *args)
-      end
-
-      def visit_key(node, opts = {})
-        name, predicate = node
-        visit(predicate, opts.merge(name: name))
-      end
-
-      def visit_val(node, opts = {})
-        visit(node, opts)
-      end
-
-      def visit_predicate(node, base_opts = {})
-        predicate, args = node
-
-        input, rule, name, val_type = base_opts
-          .values_at(:input, :rule, :name, :val_type)
-
-        val_type ||= input.class if input
-
-        lookup_options = base_opts.merge(
-          val_type: val_type,
-          arg_type: args.size > 0 && args[0][1].class,
-          message_type: message_type,
-          locale: locale
-        )
-
-        tokens = options_for(predicate, args)
-        template = messages[predicate, lookup_options.merge(tokens)]
-
-        name ||= tokens[:name]
-        rule ||= (name || tokens[:name])
-
-        unless template
-          raise MissingMessageError.new("message for #{predicate} was not found")
-        end
-
-        rule_name =
-          if rule.is_a?(Symbol)
-            messages.rule(rule, lookup_options) || rule
-          else
-            rule
-          end
-
-        text =
-          if full?
-            "#{rule_name || tokens[:name]} #{template % tokens}"
-          else
-            template % tokens
-          end
-
-        if name.is_a?(Array)
-          path = name
-        else
-          path = base_opts.fetch(:path, Array(name))
-          path = path + [name] unless path.last == name || name.nil?
-        end
-
-        is_each = base_opts[:each] == true
-
-        *arg_vals, _ = args.map(&:last)
-        message_class.new(predicate, path, text, args: arg_vals, rule: rule, each: is_each)
-      end
-
-      def visit_each(node, opts = {})
-        node.map { |el| visit(el, opts) }
-      end
-
-      def visit_set(node, opts = {})
-        node.map { |input| visit(input, opts) }
-      end
-
-      def visit_el(node, opts = {})
-        idx, el = node
-        visit(el, opts.merge(path: opts[:path] + [idx]))
+      def visit_schema(node, opts = {})
+        visit(node)
       end
 
       def visit_check(node, opts = {})
         path, other = node
         visit(other, opts.merge(path: Array(path)))
-      end
-
-      def dump_messages(hash)
-        hash.each_with_object({}) do |(key, val), res|
-          res[key] =
-            case val
-            when Hash then dump_messages(val)
-            when Array then val.map(&:to_s)
-            end
-        end
-      end
-
-      def options_for(predicate, args)
-        meth = :"options_for_#{predicate}"
-
-        defaults = Hash[args]
-
-        if respond_to?(meth)
-          defaults.merge!(__send__(meth, defaults))
-        end
-
-        defaults
-      end
-
-      def options_for_inclusion?(args)
-        warn 'inclusion is deprecated - use included_in instead.'
-        options_for_included_in?(args)
-      end
-
-      def options_for_exclusion?(args)
-        warn 'exclusion is deprecated - use excluded_from instead.'
-        options_for_excluded_from?(args)
-      end
-
-      def options_for_excluded_from?(args)
-        { list: args[:list].join(', ') }
-      end
-
-      def options_for_included_in?(args)
-        { list: args[:list].join(', ') }
-      end
-
-      def options_for_size?(args)
-        size = args[:size]
-
-        if size.is_a?(Range)
-          { left: size.first, right: size.last }
-        else
-          args
-        end
       end
     end
   end
